bioinfo/bowtie: fixed symlink from index to genome sequence
bowtie-index pipeline would add a symlink to the fasta sequence just
next to the index, so that bowtie doesn't have to reconstruct it
before aligning. However this is a bug, as the fasta pipeline might be
deleted in the cache, which breaks the symlink. The current fix is to
copy the fasta file instead of symlinking it but maybe that's not
worth it?