guizmin_table: richer typing of file formats
the type ['a Guizmin_table.format] is now reacher as it indicates the
possible presence of comments and header in addition to the format of
a row. This enabled the static detection of a bug in test_mouse_ES,
namely that you can't use macs xls output as bed files.

Note that the functions to read tables in [Guizmin_table] are not yet
able to ignore comments.